made changes from pull request reviews & tweaked order history page to show name
