+++ client/components/detailed-book.js
@@ -4,6 +4,7 @@ import {fetchReviewsForBook} from '../store/reviews.js'
 import Reviews from './reviews'
 import {withRouter} from 'react-router-dom'
 import {Container, Image, Divider} from 'semantic-ui-react'
+import {prettyDollar} from '../utils'
 import WriteReview from './write-review'
 
 export class DetailedBook extends Component {
@@ -31,9 +32,7 @@ export class DetailedBook extends Component {
             <Image size="medium" src={`/${selectedBook.imgUrl}`} centered />
             <h5>
               {/* Book price is an integer in db, so we need to reformat it as a price */}
-              {`$${selectedBook.price
-                .toString()
-                .slice(0, -2)}.${selectedBook.price.toString().slice(-2)}`}
+              {prettyDollar(selectedBook.price)}
             </h5>
             {!selectedBook.quantity && <h5>Out of stock</h5>}
             <h6>{`Format: ${selectedBook.editionType}`}</h6>
