+++ client/components/users-order-history.js
@@ -1,6 +1,7 @@
 import React, {Component} from 'react'
 import {Segment, Dimmer, Loader, Image, Item, Header} from 'semantic-ui-react'
 import {fetchUsersOrdersFromServer} from '../store/orders'
+import {prettyDate, prettyDollar} from '../utils'
 import {Link} from 'react-router-dom'
 import {connect} from 'react-redux'
 
@@ -18,42 +19,6 @@ const mapStateToProps = state => {
   }
 }
 
-export const prettyDollar = aNum => {
-  return '$' + aNum.toString().slice(0, -2) + '.' + aNum.toString().slice(-2)
-}
-
-export const prettyDate = aDate => {
-  const monthNames = [
-    'Jan',
-    'Feb',
-    'March',
-    'April',
-    'May',
-    'June',
-    'July',
-    'Aug',
-    'Sep',
-    'Oct',
-    'Nov',
-    'Dec'
-  ]
-  let day = ''
-  let monthIndex = 0
-  const startDay = aDate.slice(8, 10)
-  if (startDay.slice(0, 1) === '0') {
-    day = startDay.slice(1)
-  } else {
-    day = startDay
-  }
-  const startMonth = aDate.slice(5, 7)
-  if (startMonth.slice(0, 1) === '0') {
-    monthIndex = Number(startMonth.slice(1))
-  }
-  const year = aDate.slice(0, 4)
-
-  return day + ' ' + monthNames[monthIndex] + ' ' + year
-}
-
 class UsersOrders extends Component {
   componentDidMount() {
     const accountId = this.props.match.params.accountId
@@ -69,11 +34,10 @@ class UsersOrders extends Component {
             <Loader>Loading</Loader>
           </Dimmer>
 
-          <Image src="http://www.clker.com/cliparts/R/w/q/4/j/l/book.svg" />
+          <Image src="/img/book" />
         </Segment>
       )
     }
-    console.log('user object', this.props.user)
     if (
       this.props.user.id === Number(this.props.match.params.accountId) ||
       this.props.user.isAdmin
@@ -81,7 +45,7 @@ class UsersOrders extends Component {
       return (
         <div>
           <Header as="h2" attached="top">
-            Your Orders{' '}
+            Welcome Back {list[0] && list[0].user.firstName ? list[0].user.firstName + '!' : '!'}
           </Header>
           <Item.Group link>
             {list.map(order => {
